refactor(ProductController,GlobalExceptionHandler): move 'handleValidationException()' from controller to global exception handler
