+++ src/main/kotlin/ecommerce/GlobalExceptionHandler.kt
@@ -4,6 +4,7 @@ import ecommerce.exception.InternalServerErrorException
 import ecommerce.exception.NotFoundException
 import org.springframework.dao.DataAccessException
 import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
 
@@ -31,5 +32,16 @@ class GlobalExceptionHandler {
     fun handlerIllegalStateException(e: Exception): ResponseEntity<Void> {
         println("IllegalStateException occurred: " + e.message)
         return ResponseEntity.internalServerError().build()
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException::class)
+    fun handleValidationException(e: MethodArgumentNotValidException): ResponseEntity<Map<String, Any>> {
+        val errors =
+            e.bindingResult.fieldErrors.associate { error ->
+                error.field to (error.defaultMessage ?: "Invalid value")
+            }
+        val errorBody = mapOf("errors" to errors)
+        println("MethodArgumentNotValidException occurred: $errorBody")
+        return ResponseEntity.badRequest().body(errorBody)
     }
 }
